Renaming Layers.OneHanded and Layers.TwoHanded
These are being renamed so that their meaning is more clear; these layers do not represent OneHanded or TwoHanded items properly, and they do not represent Left or Right hands properly. They are arbitrary item slots for the hands. The rendering and client logic previously looked at these values from conflicting perspectives, and as previously named, the meaning conflicted with the data in practice.

Adding a Dress Queue so that equipping one item is delayed by first un-equipping the conflicting items.

Adding a new CSV file for specifying extended item data that is currently not supported by the existing statics or network protocol.

Adding extended data support on Item

Adding a new find item function for looking at the paperdoll hands.

Integration of the UOS scripting command "togglehands"

1. adding a new function ClearEquipped to unequip and unset the cached toggle item for that hand
2. updating the ClearHands command to use the new function

Adding a StaticLayer property so that it is easier to reference an Item's given Layer; previously needed to be provided externally.

Shortening the dress queue delay so that dressing feels more responsive.

Integration of the "equipitem" UOS command.